App splitted and added time differential
App has been splitted into several classes and files, resulting in a menu like structure. Furthermore, the way of keeping the track saved has been changed. The List with points is no longer of the type PointF, but of a floatarray containing x on position 1, y on position 2, and time differential since last start on position 3.