Code and decode a track to and from string
Tracks can now be converted to strings, and strings that have been generated that way can be converted back to a track! This function is needed to pass tracks between activities and to save / load tracks. It has now been used already to display a fake track when the user hasnt run yet.